requirements.txt
aiofiles==24.1.0
arq==0.26.0
databases[postgresql]==0.9.0
fastapi==0.111.1
gunicorn==22.0.0
pandas==2.2.2
python-dotenv==1.0.1
watchfiles==0.22.0
Jinja2==3.1.4